separate angular directive, move to render complete to kibana_utils
