+++ src/legacy/ui/public/render_complete/index.ts
@@ -17,19 +17,4 @@
  * under the License.
  */
 
-const dispatchCustomEvent = (el: HTMLElement, eventName: string) => {
-  // we're using the native events so that we aren't tied to the jQuery custom events,
-  // otherwise we have to use jQuery(element).on(...) because jQuery's events sit on top
-  // of the native events per https://github.com/jquery/jquery/issues/2476
-  el.dispatchEvent(new CustomEvent(eventName, { bubbles: true }));
-};
-
-export function dispatchRenderComplete(el: HTMLElement) {
-  dispatchCustomEvent(el, 'renderComplete');
-}
-
-export function dispatchRenderStart(el: HTMLElement) {
-  dispatchCustomEvent(el, 'renderStart');
-}
-
-export * from './render_complete_helper';
+import './directive';
