+++ src/legacy/ui/public/visualize/components/visualization_chart.tsx
@@ -22,7 +22,10 @@ import * as Rx from 'rxjs';
 import { debounceTime, filter, share, switchMap, tap } from 'rxjs/operators';
 
 import { PersistedState } from '../../persisted_state';
-import { dispatchRenderComplete, dispatchRenderStart } from '../../render_complete';
+import {
+  dispatchRenderComplete,
+  dispatchRenderStart,
+} from '../../../../plugins/kibana_utils/public';
 import { ResizeChecker } from '../../resize_checker';
 import { Vis, VisualizationController } from '../../vis';
 import { getUpdateStatus } from '../../vis/update_status';
