+++ src/legacy/ui/public/visualize/components/visualization_requesterror.tsx
@@ -20,7 +20,7 @@
 import { EuiIcon, EuiSpacer, EuiText } from '@elastic/eui';
 import React from 'react';
 import { SearchError } from 'ui/courier';
-import { dispatchRenderComplete } from '../../render_complete';
+import { dispatchRenderComplete } from '../../../../plugins/kibana_utils/public';
 
 interface VisualizationRequestErrorProps {
   onInit?: () => void;
