+++ src/legacy/ui/public/visualize/loader/embedded_visualize_handler.ts
@@ -29,7 +29,7 @@ import { Inspector } from '../../inspector';
 import { Adapters } from '../../inspector/types';
 import { PersistedState } from '../../persisted_state';
 import { IPrivate } from '../../private';
-import { RenderCompleteHelper } from '../../render_complete';
+import { RenderCompleteHelper } from '../../../../plugins/kibana_utils/public';
 import { AppState } from '../../state_management/app_state';
 import { timefilter } from '../../timefilter';
 import { RequestHandlerParams, Vis } from '../../vis';
