+++ src/plugins/kibana_utils/public/render_complete/render_complete_helper.ts
